Move Main component, rename teck-stack.md
Move Main.tsx to: client/src/components
- update import paths as needed

Small update to App.tsx

Rename tech-stack.md to TECHSTACK.md
