+++ client/src/App.tsx
@@ -1,7 +1,7 @@
 import React, { memo } from 'react';
 
 import Body from './components/Body';
-import Main from './containers/Main';
+import Main from './components/Main';
 import Nav from './containers/Nav/Nav';
 
 
@@ -9,7 +9,7 @@ const App = memo(props => {
   return (
     <Body>
       <Nav/>
-      <Main></Main>
+      <Main/>
     </Body>
   );
 })
